refactor: only keep relevant modules and group related ones together
