+++ crates/primitives/src/params/connectors.rs
@@ -1,4 +1,5 @@
 //! Params related to the bridge tx graph connectors;
+use std::time::Duration;
 
 pub const NUM_PKS_A256_PER_CONNECTOR: usize = 7;
 pub const NUM_PKS_A256: usize = 42;
@@ -32,3 +33,19 @@ pub const TOTAL_VALUES: usize = NUM_ASSERT_DATA_TX1 * NUM_ASSERT_DATA_TX1_A256_P
 
 // compile time to check to ensure that the numbers are sound.
 const _: [(); 0] = [(); (TOTAL_VALUES - TOTAL_CONNECTORS)];
+
+// FIXME: Move these to configurable params
+
+pub const BLOCK_TIME: Duration = Duration::from_secs(30);
+
+pub const SUPERBLOCK_MEASUREMENT_PERIOD: u32 = 100; // blocks
+
+pub const TS_COMMITMENT_MARGIN: u32 = 288; // 2 days' worth of blocks in mainnet
+
+pub const PAYOUT_OPTIMISTIC_TIMELOCK: u32 = 3000;
+
+const _: u32 =
+    PAYOUT_OPTIMISTIC_TIMELOCK - (SUPERBLOCK_MEASUREMENT_PERIOD + TS_COMMITMENT_MARGIN + 100); // 100
+                                                                                               // is slack
+
+pub const PAYOUT_TIMELOCK: u32 = 288; // 2 day's worth of blocks in mainnet
